Move repo tasks store files to repo-tasks-store dir
