+++ extensions/ql-vscode/src/variant-analysis/repo-tasks-store/repo-task-data-types.ts
