+++ extensions/ql-vscode/src/variant-analysis/repo-tasks-store/repo-task-to-data-mapper.ts
